MB-6161 Consider write queue size and disk usage for compaction
This change allows us to consider the disk write queue size and
current disk usage in determining if compaction is required or
can be skipped.

If the disk write queue size is above its max cap and the current
disk usage is still below 2/3 of disk space, we skip the
compaction temporarily.

On the other hand, if the current disk usage is above 90% of
disk space, we schedule the compaction as long as the
fragmentation ratio is above 2%.

In other cases, we schedule the compaction for a given vbucket
if its fragmentation is above the threshold (50% by default).

Change-Id: I54219b4d314a3d9b2e5163530f719d90f29801c6